requirements.txt
asgiref==3.2.3
Django==3.0.1
django-nested-inline==0.4.0
mysqlclient==1.4.6
Pillow==6.2.1
pytz==2019.3
sqlparse==0.3.0
